feat: tried to post a new habit

+++ src/main/java/com/isadora/habittracker/domain/HabitResponse.java
@@ -1,7 +1,7 @@
 package com.isadora.habittracker.domain;
 
 /* This is an example */
-public record HabitResponse(int habitId, String habitName, int difficultyPoints) {
+public record HabitResponse(long habitId, String habitName, int difficultyPoints) {
 
     public static HabitResponse of(Habit habit) {
         return new HabitResponse(habit.getId(), habit.getHabitName(), habit.getDifficultyPoints());
